+++ test/dynamo/test_repros.py
@@ -49,9 +49,6 @@ from torch.testing._internal.common_cuda import (
     TEST_CUDA,
 )
 from torch.testing._internal.common_device_type import instantiate_device_type_tests
-from torch.testing._internal.common_distributed import (
-    requires_nccl_version_less_or_equal,
-)
 from torch.testing._internal.common_utils import (
     disable_translation_validation_if_dynamic_shapes,
     instantiate_parametrized_tests,
@@ -6371,8 +6368,6 @@ def forward(self, s0 : torch.SymInt, s1 : torch.SymInt, L_x_ : torch.Tensor):
 
     @unittest.skipIf(not TEST_CUDA, "test requires CUDA")
     @unittest.skipIf(not dist.is_available(), "test requires distributed")
-    # Failing with ncc update 2.25.1 : https://github.com/pytorch/pytorch/issues/147141
-    @requires_nccl_version_less_or_equal((2, 21), "Failing with nccl 2.25.1")
     def test_ddp_checkpoint(self):
         # https://github.com/pytorch/pytorch/issues/144035
         DIM = 256
