+++ torch/testing/_internal/common_distributed.py
@@ -330,17 +330,6 @@ def requires_nccl_version(version, msg):
             f"Requires NCCL version greater than or equal to: {version}, found: {torch.cuda.nccl.version()}, reason: {msg}",
         )
 
-def requires_nccl_version_less_or_equal(version, msg):
-    if not c10d.is_nccl_available():
-        return skip_but_pass_in_sandcastle(
-            "c10d was not compiled with the NCCL backend",
-        )
-    else:
-        return skip_but_pass_in_sandcastle_if(
-            torch.cuda.nccl.version() >= version,
-            f"Requires NCCL version less then than or equal to: {version}, found: {torch.cuda.nccl.version()}, reason: {msg}",
-        )
-
 def requires_nccl():
     return skip_but_pass_in_sandcastle_if(
         not c10d.is_nccl_available(),
